REFACTOR: use lower case package name
