+++ src/main/java/segelzwerg/sporttooolbox/iunits/speed/SpeedFactory.java
@@ -1,4 +1,4 @@
-package segelzwerg.sporttooolbox.IUnits.speed;
+package segelzwerg.sporttooolbox.iunits.speed;
 
 import java.util.Objects;
