+++ src/main/java/segelzwerg/sporttooolbox/web/pace/PaceService.java
@@ -1,12 +1,12 @@
 package segelzwerg.sporttooolbox.web.pace;
 
 import org.springframework.stereotype.Component;
-import segelzwerg.sporttooolbox.IUnits.Time;
-import segelzwerg.sporttooolbox.IUnits.pace.Pace;
 import segelzwerg.sporttooolbox.calculators.SpeedCalculator;
 import segelzwerg.sporttooolbox.calculators.SpeedCalculatorFactory;
 import segelzwerg.sporttooolbox.calculators.TimeCalculator;
 import segelzwerg.sporttooolbox.calculators.TimeCalculatorFactory;
+import segelzwerg.sporttooolbox.iunits.Time;
+import segelzwerg.sporttooolbox.iunits.pace.Pace;
 import segelzwerg.sporttooolbox.web.speed.SpeedForm;
 
 import java.util.ArrayList;
