+++ src/test/java/segelzwerg/sporttooolbox/web/pace/PacePresenterTest.java
@@ -2,8 +2,8 @@ package segelzwerg.sporttooolbox.web.pace;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import segelzwerg.sporttooolbox.IUnits.pace.MinutesPerHundredMeters;
-import segelzwerg.sporttooolbox.IUnits.pace.MinutesPerKilometer;
+import segelzwerg.sporttooolbox.iunits.pace.MinutesPerHundredMeters;
+import segelzwerg.sporttooolbox.iunits.pace.MinutesPerKilometer;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
